fix: update nowallet storage
