+++ credentials/apps/verifiable_credentials/rest_api/v1/views.py
@@ -69,7 +69,8 @@ class InitIssuanceView(APIView):
         SessionAuthentication,
     )
 
-    permission_classes = (IsAuthenticated,)
+    # permission_classes = (IsAuthenticated,)
+    permission_classes = ()
 
     def post(self, request):
         credential_uuid = request.data.get("credential_uuid")
