+++ credentials/apps/verifiable_credentials/settings.py
@@ -26,7 +26,7 @@ DEFAULTS = {
     ],
     "DEFAULT_ISSUER_DID": None,
     "DEFAULT_ISSUER_KEY": None,
-    "DEFAULT_ISSUANCE_REQUEST_SERIALIZER": "credentials.apps.verifiable_credentials.issuance.IssuanceLineSerializer",
+    "DEFAULT_ISSUANCE_REQUEST_SERIALIZER": "credentials.apps.verifiable_credentials.serializers.IssuanceLineSerializer",
 }
 
 # List of settings that may be in string import notation:
